some fixes
1) Fixed bullet debt with use of Int16s instead of Int8s
2) Fixed scroll-way weapon changing, now with loopback

+++ client/src/deserialisers.ts
@@ -40,7 +40,7 @@ function _getAmmos(stream: IslandrBitStream): number[] {
     const ammos: number[] = []
     const size = stream.readInt8()
     for (let ii = 0; ii < size; ii++) {
-        const numofammo = stream.readInt8()
+        const numofammo = stream.readInt16()
         ammos[ii] = numofammo
     }
     return ammos
